Refactor some utils into services
