+++ scripts/resizeIcon.ts
@@ -1,5 +1,5 @@
-import IconResizer from "../utils/IconResizer";
-import Prompter from "../utils/Prompter";
+import IconResizer from "../services/IconResizer";
+import Prompter from "../services/Prompter";
 
 (async () => {
   const { iconToResize } = await Prompter.forIconResizing();
